+++ README.md
@@ -260,7 +260,9 @@ dependencies {
 ## STEP 5. 쿠키를 이용한 로그인
 
 ### 1. Cookie
-- 하이퍼 텍스트의 기록서(HTTP)의 일종으로서 인터넷 사용자가 어떠한 웹사이트를 방문할 경우 그 사이트가 사용하고 있는 서버를 통해 인터넷 사용자의 컴퓨터에 설치되는 작은 기록 정보 파일
+
+#### 1) 개념
+- 클라이언트(로컬)에 저장되는 키와 값이 들어있는 작은 데이터 파일
 - 쿠키에 담긴 정보는 인터넷 사용자가 같은 웹사이트를 방문할 때마다 읽히고 수시로 새로운 정보로 바뀜
 - 소프트웨어가 아니므로 컴퓨터 내에서 프로그램처럼 실행될 수 없으며 바이러스를 옮길 수도, 악성코드를 설치할 수도 없음
 - 단, 스파이웨어를 통해 유저의 브라우징 행동을 추적하는데에 사용될 수 있음
@@ -271,7 +273,16 @@ dependencies {
 - 클라이언트에서 직접 쿠키에 데이터를 저장 할 수도 있음
 - 유저의 별다른 컨트롤 없이, 요청 시 자동으로 request header에 넣어서 서버에 전송
 - 만료기간을 설정할 수 있으며, 만료기간 전까지 브라우저가 종료되어도 삭제되지 않음
+- Expires 또는 Max-Age를 지정하지 않으면, Session 쿠키로 간주하여 브라우저 종료시 쿠키가 소멸
 
+#### 2) 존재 이유
+1. HTTP의 Connectionless
+   - HTTP는 TCP 연결을 맺고 요청을 보내면 서버는 응답을 보내고 연결이 끊어짐
+2. HTTP의 Stateless
+   - HTTP는 상태를 따로 저장하지 않음 
+   - 연결이 끊어지는 순간 모든 상태 정보가 사라지게 됨
+
+#### 3) 속성
 1. key-value
    - value에 실제 데이터가 담겨져 있음
    - key값으로 쿠키의 value값을 조회, 갱신, 삭제 가능
@@ -282,9 +293,26 @@ dependencies {
    - 해당 도메인 및 하위 도메인에서만 쿠키에 접근 가능
 4. expires
    - 쿠키의 만료기간
-   - 만료기간이 지나면 쿠키 자동 삭제
-5. secure
+   - 만료기간이 지나면 쿠키 자동 삭제 (절대적)
+   - GMT 형식
+5. Max-Age
+   - 현재 시간을 기준으로 쿠키를 유지할 기간 지정 (상대적)
+6. secure
    - secure 속성을 설정하면 HTTPS 통신이 아닌경우 서버로 쿠키를 전송하지 않음
+
+### 2. Session
+- 일정 기간 동안 같은 클라이언트로부터 들어오는 일련의 요구를 하나의 상태로 보고 
+그 상태를 일정하게 유지시키는 기술
+- 쿠키를 기반으로 하지만 쿠키와 다르게 서버 측에서 저장하고 관리
+- 세션 ID를 이용해 클라이언트를 구분
+
+### 3. 쿠키와 세션의 차이
+- 쿠키는 클라이언트에 저장 vs. 세션은 서버에 저장
+- 세션은 서버에 저장되므로 서버의 자원을 사용
+- 서버에 요청을 보내는 사용자가 많을 경우 세션은 부하가 심할 수 있음
+- 보안면에서 세션이 더 우수
+- 쿠키는 만료시간 동안 파일로 저장되므로 브라우저를 종료해도 정보가 남아있음
+- 세션은 브라우저가 종료되면 만료시간에 상관없이 삭제됨
 
 -----------------------------------------------
 ## 웹 서버 2주차 추가 공부
