+++ src/main/java/dto/request/HTTPRequestDto.java
@@ -1,4 +1,4 @@
-package dto;
+package dto.request;
 
 import java.util.HashMap;
